destroy_datasets.py now reads PIDs from given CSV or TXT file

+++ destroy_datasets.py
@@ -1,5 +1,7 @@
 # Destroys a given list of datasets
 
+from csv import DictReader
+import os
 import urllib.request
 
 # Dataverse repository URL, e.g. https://demo.dataverse.org
@@ -11,32 +13,58 @@ apikey = ''
 # Text file containing PIDs of datasets to be destroyed, e.g. /Users/user/Desktop/dois.txt
 file = ''
 
-dataset_pids = open(file)
-
 destroyed_datasets = []
 not_destroyed_datasets = []
 
+file_name = os.path.basename(file)
+
 print('Trying to destroy datasets...')
 
-for dataset_pid in dataset_pids:
-    dataset_pid = dataset_pid.rstrip()
-    url = '%s/api/datasets/:persistentId/destroy/?persistentId=%s' % (server, dataset_pid)
+if '.txt' in file_name:
+    dataset_pids = open(file)
+    for dataset_pid in dataset_pids:
+        dataset_pid = dataset_pid.rstrip()
+        url = '%s/api/datasets/:persistentId/destroy/?persistentId=%s' % (server, dataset_pid)
 
-    headers = {
-        'X-Dataverse-key': apikey}
+        headers = {
+            'X-Dataverse-key': apikey}
 
-    req = urllib.request.Request(
-        url=url,
-        headers=headers,
-        method='DELETE')
+        req = urllib.request.Request(
+            url=url,
+            headers=headers,
+            method='DELETE')
 
-    try:
-        response = urllib.request.urlopen(req)
-        print('%s destroyed' % (dataset_pid))
-        destroyed_datasets.append(dataset_pid)
-    except Exception:
-        print('Could not destroy %s' % (dataset_pid))
-        not_destroyed_datasets.append(dataset_pid)
+        try:
+            response = urllib.request.urlopen(req)
+            print('%s destroyed' % (dataset_pid))
+            destroyed_datasets.append(dataset_pid)
+        except Exception:
+            print('Could not destroy %s' % (dataset_pid))
+            not_destroyed_datasets.append(dataset_pid)
+
+else:
+    if '.csv' in file_name:
+        with open(file, mode='r', encoding='utf-8') as f:
+            csv_dict_reader = DictReader(f, delimiter=',')
+            for row in csv_dict_reader:
+                dataset_pid = row['persistent_id'].rstrip()
+                url = '%s/api/datasets/:persistentId/destroy/?persistentId=%s' % (server, dataset_pid)
+
+                headers = {
+                    'X-Dataverse-key': apikey}
+
+                req = urllib.request.Request(
+                    url=url,
+                    headers=headers,
+                    method='DELETE')
+
+                try:
+                    response = urllib.request.urlopen(req)
+                    print('%s destroyed' % (dataset_pid))
+                    destroyed_datasets.append(dataset_pid)
+                except Exception:
+                    print('Could not destroy %s' % (dataset_pid))
+                    not_destroyed_datasets.append(dataset_pid)
 
 print('\nDatasets destroyed: %s' % (len(destroyed_datasets)))
 if destroyed_datasets:
